requirements.txt
arrow==1.4.0
canvasapi==3.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-frontmatter==1.1.0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
six==1.17.0
tzdata==2025.3
urllib3==2.6.2
